reduce class
