+++ src/main/kotlin/com/improve_future/harmonica/plugin/AbstractMigrationTask.kt
@@ -21,8 +21,12 @@ package com.improve_future.harmonica.plugin
 import com.improve_future.harmonica.core.*
 import com.improve_future.harmonica.service.VersionService
 import org.gradle.api.tasks.Input
+import org.jetbrains.kotlin.script.jsr223.KotlinJsr223JvmLocalScriptEngine
+import java.io.File
+import java.nio.file.Paths
+import javax.script.ScriptEngineManager
 
-abstract class AbstractMigrationTask : AbstractHarmonicaTask() {
+abstract class AbstractMigrationTask : AbstractTask() {
     /** The table name to store executed migration version IDs. */
     private val migrationTableName: String = "harmonica_migration"
     protected val versionService: VersionService
@@ -38,11 +42,24 @@ abstract class AbstractMigrationTask : AbstractHarmonicaTask() {
         return engine.eval(removePackageStatement(script)) as AbstractMigration
     }
 
+    private fun findConfigFile(): File {
+        return Paths.get(directoryPath, "config", "$env.kts").toFile()
+    }
+
+    fun loadConfigFile(): DbConfig {
+        return engine.eval(findConfigFile().readText()) as DbConfig
+    }
+
     protected fun createConnection(): Connection {
         return Connection(loadConfigFile())
     }
 
     protected companion object {
+        val engine: KotlinJsr223JvmLocalScriptEngine by lazy {
+            ScriptEngineManager().getEngineByName("kotlin") as
+                    KotlinJsr223JvmLocalScriptEngine
+        }
+
         protected fun removePackageStatement(script: String) =
             script.replace(Regex("^\\s*package\\s+.+"), "")
     }
